Build: RunClangTidy.py, the clang-tidy gate over the whole tree (m0-foundation/T5)
Every ClCompile item of every project the solution lists runs through
the pinned clang-tidy with clang's MSVC driver and the switches ADR-001
fixes, the include directories and definitions read from the .vcxproj,
and the repository's .clang-tidy; suites get the framework header's
directory, found under VCINSTALLDIR. Units run in parallel, the output
is printed unchanged, and any finding fails. INCLUDE unset stops the
run with one line; the version found is printed and a mismatch with
the workflow's pin is a warning. --dry-run prints the commands.

The first pass, run locally with the pinned binary over Core and its
suite, found two defects: the log's namespace-scope std::ofstream,
whose constructor could throw before main, is now a function-local
static, and the rotation default argument multiplies in the wide type.

.clang-tidy no longer describes either checker as unwritten, its
example command drops the macro-family defines that WarningsAsErrors
would make fatal, and the vendored Client/d3dx12.h joins Resource.h in
ExcludeHeaderFilterRegex. AGENTS.md says all three checkers exist.

+++ Core/Log.h
@@ -28,7 +28,8 @@ class Log
 public:
   /// Opens the file, creating its directory; a file already larger than _rotateAtBytes is
   /// renamed to <name>.1 first, replacing an older .1, so a log never grows without bound.
-  [[nodiscard]] static bool Open(const std::filesystem::path& _file, std::uintmax_t _rotateAtBytes = 4u * 1024u * 1024u);
+  [[nodiscard]] static bool Open(const std::filesystem::path& _file,
+                                 std::uintmax_t _rotateAtBytes = static_cast<std::uintmax_t>(4) * 1024 * 1024);
   static void Close();
 
   /// Lines below the minimum are dropped; Info by default.
